requirements.txt
appdirs==1.4.0
blinker==1.4
boto3==1.4.4
botocore==1.5.31
coreapi==2.3.0
coreschema==0.0.4
crcmod==1.7
Django==1.10.5
django-filter==1.0.1
django-ipware==1.1.6
django-rest-framework-mongoengine==3.3.1
django-rest-swagger==2.1.1
django-simple-captcha==0.5.3
djangorestframework==3.5.4
djangorestframework-gis==0.11
djangorestframework-jwt==1.9.0
docutils==0.13.1
httplib2==0.10.3
itypes==1.1.0
Jinja2==2.9.5
jmespath==0.9.2
Markdown==2.6.8
MarkupSafe==1.0
mongoengine==0.11.0
olefile==0.44
openapi-codec==1.3.1
oss2==2.2.3
packaging==16.8
Pillow==4.0.0
py==1.4.32
pydenticon==0.3
Pygments==2.2.0
PyJWT==1.4.2
pymongo==3.4.0
pyparsing==2.1.10
pytest==3.0.6
python-dateutil==2.6.0
python-Levenshtein==0.12.0
requests==2.12.5
s3transfer==0.1.10
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
